Update meshes to use 32 bit uints

+++ src/core/resource/IResourceFactory.hpp
@@ -19,7 +19,7 @@ public:
    virtual std::unique_ptr<ITexture> CreateRenderTarget(const uint32_t width, const uint32_t height,
                                                         const ITexture::Format format = ITexture::Format::RGBA8, const uint32_t samples = 1) = 0;
    // Mesh creation methods
-   virtual std::unique_ptr<IMesh> CreateMesh(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices) = 0;
+   virtual std::unique_ptr<IMesh> CreateMesh(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices) = 0;
    virtual std::unique_ptr<IMesh> CreateMeshFromFile(const std::string& filepath) = 0;
 };
