+++ src/core/resource/ResourceManager.cpp
@@ -77,7 +77,7 @@ TextureHandle ResourceManager::CreateRenderTarget(const std::string& name, const
 }
 
 MeshHandle ResourceManager::LoadMesh(const std::string& name, const std::vector<Vertex>& vertices,
-                                     const std::vector<uint16_t>& indices) {
+                                     const std::vector<uint32_t>& indices) {
    auto mesh = m_factory->CreateMesh(vertices, indices);
    if (!mesh || !mesh->IsValid()) {
       return MeshHandle();
