+++ src/gl/resource/GLMesh.hpp
@@ -10,7 +10,7 @@
 
 class GLMesh : public IMesh {
 public:
-   GLMesh(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices);
+   GLMesh(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices);
    ~GLMesh();
 
    GLMesh(const GLMesh&) = delete;
